requirements.txt
python-dotenv==1.0.0
python=3.10
openai==1.12.0
pymysql==1.1.0